Add Sprig all-in-one agent binary
