+++ crates/sprig/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "sprig"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "All-in-one Sprout ACP harness, agent, and developer MCP"
+
+[[bin]]
+name = "sprig"
+path = "src/main.rs"
+
+[dependencies]
+sprout-acp = { path = "../sprout-acp" }
+sprout-agent = { path = "../sprout-agent" }
+sprout-dev-mcp = { path = "../sprout-dev-mcp" }
